+++ framework/validators/EachValidator.php
@@ -128,6 +128,9 @@ class EachValidator extends Validator
      */
     public function validateAttribute($model, $attribute)
     {
+        // PHP 7.4 Typed properties not initialized
+        if(!isset($model->$attribute)) return;
+
         $value = $model->$attribute;
         if (!is_array($value) && !$value instanceof \ArrayAccess) {
             $this->addError($model, $attribute, $this->message, []);