+++ tests/framework/validators/EachValidatorTest.php
@@ -12,6 +12,7 @@ use yii\validators\EachValidator;
 use yiiunit\data\base\ArrayAccessObject;
 use yiiunit\data\base\TraversableObject;
 use yiiunit\data\validators\models\FakedValidationModel;
+use yiiunit\data\validators\models\ValidatorTestTypedPropModel;
 use yiiunit\TestCase;
 
 /**
@@ -199,5 +200,19 @@ class EachValidatorTest extends TestCase
         $this->assertFalse($model->hasErrors('array'));
 
         $this->assertTrue($validator->validate($model->attr_array));
+    }
+
+    public function testTypedProperties()
+    {
+        if (PHP_VERSION_ID < 70400) {
+            $this->markTestSkipped('Can not be tested on PHP < 7.4');
+            return;
+        }
+
+        $model = new ValidatorTestTypedPropModel();
+
+        $validator = new EachValidator(['rule' => ['boolean']]);
+        $validator->validateAttribute($model, 'arrayTypedProperty');
+        $this->assertFalse($model->hasErrors('array'));
     }
 }